Working on registry

+++ assets/js/pages/Registries.js
@@ -1,9 +1,13 @@
 import React, {Component} from 'react'
+import { Link } from 'react-router'
 import ReactTooltip from 'react-tooltip'
+import Btn from './../components/Btn'
 import ContentRow from './../components/ContentRow'
 import RegistryProviderIcons from './../util/RegistryProviderIcons'
 import CenteredConfirm from './../components/CenteredConfirm'
+import ErrorMsg from './../components/ErrorMsg'
 import Loader from './../components/Loader'
+
 
 export default class Registries extends Component {
 	constructor(props) {
@@ -14,9 +18,23 @@ export default class Registries extends Component {
 		this.context.actions.listRegistries();
 	}
 	renderRegistries(){
+		if(this.context.state.registriesXHR) {
+			return (
+				<div className="PageLoader">
+					<Loader />
+				</div>	
+			);
+		}
+
+		let registries = this.context.state.registries;
+
+		if(!registries.length) {
+			return this.renderNoRegistries()
+		}
+
 		return (
 			<div className="RegistryList FlexColumn">
-				{this.context.state.registries.map(this.renderRegistryItem.bind(this))}
+				{registries.map(this.renderRegistryItem.bind(this))}
 			</div>
 		);
 	}
@@ -41,8 +59,8 @@ export default class Registries extends Component {
 				</span>
 				<span className="Pipe">|</span>
 				<span className="Actions">
-					<i className="icon icon-dis-edit" data-tip="Edit Credentials" data-for="ToolTip"/>
-					<i className="icon icon-dis-trash" data-tip="Delete Credentials" data-for="ToolTip" 
+					<i className="icon icon-dis-edit" data-tip="Edit Credentials" data-for="ToolTipTop"/>
+					<i className="icon icon-dis-trash" data-tip="Delete Credentials" data-for="ToolTipTop" 
 						onClick={() => this.context.actions.setRegistryForDelete(reg)}
 					/>
 				</span>
@@ -51,33 +69,60 @@ export default class Registries extends Component {
 			</div>
 		);
 	}
+	renderNoRegistries(){
+		return (
+			<div className="NoContent">
+				<h3>
+					No Registry Credentials Saved
+				</h3>		
+			</div>
+		);
+	}
 	renderConfirmDeleteRegistry(reg){
 		if(reg == this.context.state.registry.registrySelectedForDelete) {
+
+			if(this.context.state.registry.deleteRegistryErrorMsg) {
+				return (
+					<div className="RowPadding">
+						<ErrorMsg text={this.context.state.registry.deleteRegistryErrorMsg} 
+								  close={() => this.context.actions.setRegistryForDelete()}/>
+					</div>
+				)
+			}
 
 			if(this.context.state.registry.deleteRegistryXHR) {
 				return (
 					<Loader />
 				);
 			}
-			else {
-				return (
-					<CenteredConfirm message="Are you sure you want to delete this registry?"
-								     confirmButtonText="Delete"
-								     confirmButtonStyle={{}}
-								     onConfirm={() => this.context.actions.deleteRegistry()}
-								     onCancel={() => this.context.actions.setRegistryForDelete(null)}/>
-				);	
-			}
 			
+			return (
+				<CenteredConfirm message="Are you sure you want to delete this registry?"
+							     confirmButtonText="Delete"
+							     confirmButtonStyle={{}}
+							     onConfirm={() => this.context.actions.deleteRegistry()}
+							     onCancel={() => this.context.actions.setRegistryForDelete(null)}/>
+			);	
 		}
 	}
 	render() {
+		let registries = this.context.state.registries;
+		let rLength = registries.length;
+
 		return (
 			<div className="ContentContainer">
 				<div className="PageHeader">
 					<h2>
-						Registries
+						{rLength}&nbsp;{(rLength == 1) ? 'Registry' : 'Registries'}
 					</h2>
+					<div className="FlexRow">
+						<div className="Flex1 Column">
+							<Link to="/new-registry">
+								<Btn text="Add Registry"
+									 onClick={ () => {} } />
+							</Link>
+						</div>
+					</div>
 				</div>
 				<div>
 					{this.renderRegistries()}
